Moved JWT middleware definitions to services
