+++ app/services/auth.py
@@ -5,8 +5,37 @@ from fastapi import HTTPException
 from app.db.models.user import User
 from sqlalchemy.orm import selectinload
 import app.validators.auth as validator
+from app.config.settings import settings
 from sqlalchemy.ext.asyncio import AsyncSession
-from app.middlewares.jwt import create_access_token, create_refresh_token, decode_refresh_token
+from datetime import datetime, timedelta, timezone
+
+
+def create_access_token(data: dict, expires_delta: timedelta = None) -> str:
+    to_encode = data.copy()
+    expire = datetime.now(timezone.utc) + \
+        (expires_delta if expires_delta else timedelta(
+            minutes=settings.JWT_ACCESS_EXPIRE_MINUTES))
+    to_encode.update({"exp": expire})
+    return jwt.encode(to_encode, settings.JWT_ACCESS_SECRET_KEY, algorithm=settings.JWT_ALGORITHM)
+
+
+def create_refresh_token(data: dict, expires_delta: timedelta = None) -> str:
+    to_encode = data.copy()
+    expire = datetime.now(timezone.utc) + \
+        (expires_delta if expires_delta else timedelta(
+            days=settings.JWT_REFRESH_EXPIRE_DAYS))
+    to_encode.update({"exp": expire})
+    return jwt.encode(to_encode, settings.JWT_REFRESH_SECRET_KEY, algorithm=settings.JWT_ALGORITHM)
+
+
+def decode_access_token(token: str) -> dict:
+    return jwt.decode(token, settings.JWT_ACCESS_SECRET_KEY,
+                      algorithms=[settings.JWT_ALGORITHM], options={"verify_exp": True})
+
+
+def decode_refresh_token(token: str) -> dict:
+    return jwt.decode(token, settings.JWT_REFRESH_SECRET_KEY,
+                      algorithms=[settings.JWT_ALGORITHM], options={"verify_exp": True})
 
 
 async def authenticate_user(db: AsyncSession, data: validator.LoginRequest):
